# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: NUSA Education Operating System API
  description: |
    RESTful API for the NUSA Education Operating System.
    This API supports Kurikulum Merdeka (Indonesian National Curriculum 2024).
  version: 1.0.0
  contact:
    name: NUSA Team
    email: api@nusa.id
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.nusa.id/v1
    description: Production server
...
paths:
...
  /tp-sets/{id}:
    get:
      tags:
        - TP (Teaching Plan)
      summary: Get TP Set by ID
      description: Retrieve a specific TP Set
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TP Set retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPSetResponse'
        '404':
          description: TP Set not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tp-sets/{id}/approve:
    post:
      tags:
        - TP (Teaching Plan)
      summary: Approve TP Set
      description: Approve a TP Set
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveTPSetRequest'
      responses:
        '200':
          description: TP Set approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tps:
    post:
      tags:
        - TP (Teaching Plan)
      summary: Create a new TP
      description: Create a new Teaching Plan item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTPRequest'
      responses:
        '201':
          description: TP created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPResponse'
    get:
      tags:
        - TP (Teaching Plan)
      summary: List TPs
      description: Get a paginated list of Teaching Plans
      security:
        - bearerAuth: []
      parameters:
        - name: tp_set_id
          in: query
          description: Filter by TP Set ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [DRAFT, PENDING, APPROVED, REJECTED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: TPs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tps:
                    type: array
                    items:
                      $ref: '#/components/schemas/TPResponse'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
  /tps/{id}:
    get:
      tags:
        - TP (Teaching Plan)
      summary: Get TP by ID
      description: Retrieve a specific Teaching Plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TP retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPResponse'
        '404':
          description: TP not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - TP (Teaching Plan)
      summary: Update TP
      description: Update an existing Teaching Plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTPRequest'
      responses:
        '200':
          description: TP updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPResponse'
    delete:
      tags:
        - TP (Teaching Plan)
      summary: Delete TP
      description: Delete a Teaching Plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: TP deleted successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    # TP Schemas
...
    WorkflowStatus:
      type: string
      enum: [DRAFT, PENDING, APPROVED, REJECTED]
    GenerationSource:
      type: string
      enum: [AI_GENERATED, MANUAL]
    TPSetResponse:
      type: object
      properties:
        id:
          type: string
        cp_id:
          type: string
        cp_code:
          type: string
        cp_text:
          type: string
        version_no:
          type: integer
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        generation_source:
          $ref: '#/components/schemas/GenerationSource'
        generation_reason:
          type: string
        generated_by:
          type: string
        generated_by_name:
          type: string
        ai_generation_id:
          type: string
        approved_by:
          type: string
        approved_by_name:
          type: string
        approved_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ApproveTPSetRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 5
    TPResponse:
      type: object
      properties:
        id:
          type: string
        tp_set_id:
          type: string
        sequence_number:
          type: integer
        cp_id:
          type: string
        cp_code:
          type: string
        cp_text:
          type: string
        subject_id:
          type: string
        subject_code:
          type: string
        subject_name:
          type: string
        phase_id:
          type: string
        phase_code:
          type: string
        phase_name:
          type: string
        element_id:
          type: string
        element_code:
          type: string
        element_name:
          type: string
        subelement_id:
          type: string
        subelement_code:
          type: string
        subelement_name:
          type: string
        user_id:
          type: string
        user_name:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        title:
          type: string
        learning_objectives:
          type: object
        time_allocation:
          type: object
        prerequisites:
          type: object
        estimated_weeks:
          type: integer
        success_criteria:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTPRequest:
      type: object
      required:
        - tp_set_id
        - sequence_number
        - cp_id
        - subject_id
        - phase_id
        - element_id
        - subelement_id
        - user_id
        - learning_objectives
        - time_allocation
      properties:
        tp_set_id:
          type: string
        sequence_number:
          type: integer
          minimum: 1
        cp_id:
          type: string
        subject_id:
          type: string
        phase_id:
          type: string
        element_id:
          type: string
        subelement_id:
          type: string
        user_id:
          type: string
        title:
          type: string
        learning_objectives:
          type: object
        time_allocation:
          type: object
        prerequisites:
          type: object
        estimated_weeks:
          type: integer
        success_criteria:
          type: object
    UpdateTPRequest:
      type: object
      properties:
        title:
          type: string
        learning_objectives:
          type: object
        time_allocation:
          type: object
        prerequisites:
          type: object
        estimated_weeks:
          type: integer
        success_criteria:
          type: object
        status:
          $ref: '#/components/schemas/WorkflowStatus'

    # ATP Schemas
...